+++ genealogos-api/src/main.rs
@@ -3,12 +3,14 @@ use std::sync::{atomic, Arc};
 use genealogos::args::BomArg;
 use genealogos::backend::Backend;
 use genealogos::bom::Bom;
+use jobs::job_map::{self, garbage_collector};
 use rocket::http::Status;
 use rocket::response::{content, status};
 use rocket::serde::json::Json;
 use rocket::tokio::sync::Mutex;
 use rocket::Request;
 
+mod config;
 mod jobs;
 mod messages;
 
@@ -72,7 +74,16 @@ fn analyze(installable: &str, bom_format: Option<BomArg>) -> Result<messages::An
 
 #[rocket::launch]
 fn rocket() -> _ {
-    rocket::build()
+    let job_map = Arc::new(Mutex::new(job_map::JobHashMap::new()));
+
+    let job_map_clone = job_map.clone();
+
+    let rocket = rocket::build();
+    let figment = rocket.figment();
+
+    let config: config::Config = figment.extract().expect("Failed to load configuration");
+
+    rocket
         .attach(rocket::fairing::AdHoc::on_response("cors", |_req, resp| {
             Box::pin(async move {
                 resp.set_header(rocket::http::Header::new(
@@ -81,6 +92,16 @@ fn rocket() -> _ {
                 ));
             })
         }))
+        .attach(rocket::fairing::AdHoc::on_liftoff(
+            "garbage_collector",
+            |_| {
+                Box::pin(async move {
+                    rocket::tokio::spawn(async move {
+                        garbage_collector(job_map_clone, config.gc).await;
+                    });
+                })
+            },
+        ))
         .mount("/", rocket::routes![index])
         .mount("/api", rocket::routes![analyze])
         .register("/api", rocket::catchers![handle_errors])
@@ -89,10 +110,7 @@ fn rocket() -> _ {
             rocket::routes![jobs::create, jobs::status, jobs::result],
         )
         .register("/api/jobs/", rocket::catchers![handle_errors])
-        .manage(Arc::new(Mutex::new(std::collections::HashMap::<
-            jobs::JobId,
-            jobs::JobStatus,
-        >::new())))
+        .manage(job_map)
         .manage(atomic::AtomicU16::new(0))
 }